requirements.txt
PyYAML
tenacity
requests
geopandas
loguru
python-dotenv
tqdm
openai
tiktoken
langchain
Sphinx
beautifulsoup4
matplotlib
seaborn
tika
nbsphinx
thefuzz
ipython
unstructured
sphinx-gallery
sphinx_rtd_theme
jupyterlab